Summarize parsing logs
We used to print detailed logs if input data wasn't quite in the format
we expect. This usually happens at the beginning where GNSS data isn't
available yet but acceleration data is, and isn't a big deal.

Instead, we're only using good data and summarize at the end how much of
the total is missing.