Fix discontinuous graphs
The list of positions in a track used to contain a position for every
acceleration value, of which there are 25 per second. Each group of
those has the same timestamp, which leads to weird-looking graphs, since
there are 25 different values plotted to the same timestamp.

Position timestamps are now artificially increased by the appropriate
25th of a second so no 2 positions have the same timestamp.